FRONTEND&BACKEND/rosseti-master/requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==2.2.13
django-admin-sortable==2.2.3
django-appconf==1.0.4
django-ckeditor==6.0.0
django-colorful==1.3
django-crispy-forms==1.9.1
django-editor-ymaps==2.1.20
django-imagekit==4.0.2
django-ipware==3.0.1
django-js-asset==1.2.2
geographiclib==1.50
geopy==2.0.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
lxml==4.5.2
mysqlclient==1.4.6
pilkit==2.0
Pillow==7.1.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.0
